huge cursor agent refactoring
