+++ projects/ichi/REFACTORING_SUMMARY.md
@@ -1,0 +1,92 @@
+# ICHI Project Refactoring Summary
+
+## Overview
+
+Refactored the ICHI indexer codebase to improve readability and maintainability while preserving all functionality and comments (TODO, FIXME, XXX).
+
+## New Project Structure
+
+```
+src/
+├── types/                  # Type definitions
+│   ├── core.ts            # Core types (BalanceDelta, OnChainEvent, etc.)
+│   ├── pricing.ts         # Pricing-related types
+│   ├── adapter.ts         # Adapter interface
+│   └── index.ts
+├── cache/                  # Cache implementations
+│   ├── metadata.ts        # Redis metadata cache
+│   ├── price.ts          # Redis TimeSeries price cache
+│   └── index.ts
+├── adapters/              # Protocol adapters
+│   ├── ichi.ts           # ICHI adapter implementation
+│   ├── example-hemi.ts   # Example HEMI adapter
+│   └── index.ts
+├── engine/                # Core engine logic
+│   ├── engine.ts         # Main Engine class (streamlined)
+│   ├── pricing-engine.ts # Pricing engine and handler registry
+│   ├── asset-handlers.ts # Asset metadata handlers
+│   └── index.ts
+├── config/               # Configuration files
+│   ├── pricing.ts        # Asset pricing configurations
+│   └── ...              # Existing config files
+├── main.ts              # New clean entry point
+├── index.ts             # Project-wide exports
+├── emain.ts             # Legacy compatibility layer
+└── eprice.ts            # Legacy compatibility layer
+```
+
+## What Was Moved
+
+### From `emain.ts`:
+
+- **Types** → `types/core.ts` (BalanceDelta, OnChainEvent, etc.)
+- **Adapter interface** → `types/adapter.ts`
+- **Engine class** → `engine/engine.ts` (cleaned up)
+- **ICHI adapter** → `adapters/ichi.ts`
+- **Example adapters** → `adapters/example-hemi.ts`
+- **Feed configs** → `config/pricing.ts`
+
+### From `eprice.ts`:
+
+- **Pricing types** → `types/pricing.ts`
+- **Cache classes** → `cache/metadata.ts` and `cache/price.ts`
+- **Asset handlers** → `engine/asset-handlers.ts`
+- **Pricing engine** → `engine/pricing-engine.ts`
+
+## Backward Compatibility
+
+- `emain.ts` and `eprice.ts` now serve as compatibility layers
+- All existing imports should continue to work
+- New code should use the organized structure
+
+## Key Improvements
+
+1. **Better organization**: Related code is grouped together
+2. **Easier navigation**: Smaller, focused files
+3. **Cleaner imports**: Organized index files
+4. **Preserved comments**: All TODO, FIXME, XXX comments maintained
+5. **Maintained functionality**: No breaking changes
+
+## Usage
+
+### New way (recommended):
+
+```typescript
+import { Engine } from './engine';
+import { createIchiAdapter } from './adapters';
+import { defaultFeedConfig } from './config/pricing';
+```
+
+### Old way (still works):
+
+```typescript
+import { Engine } from './emain';
+// All existing imports continue to work
+```
+
+## Notes for Development
+
+- All existing TODO, FIXME, and XXX comments have been preserved
+- The main entry point is now `src/main.ts`
+- Use `src/index.ts` for organized project-wide exports
+- Legacy files provide backward compatibility during transition
