+++ projects/ichi/src/adapters/index.ts
@@ -1,0 +1,4 @@
+// Adapter implementations index
+
+export { createIchiAdapter } from './ichi';
+export { createHemiAdapter } from './example-hemi';
