+++ projects/ichi/src/cache/index.ts
@@ -1,0 +1,4 @@
+// Cache implementations index
+
+export { RedisMetadataCache } from './metadata';
+export { RedisTSCache } from './price';
