+++ projects/ichi/src/engine/index.ts
@@ -1,0 +1,5 @@
+// Engine module exports
+
+export { Engine } from './engine';
+export { PricingEngine, HandlerRegistry } from './pricing-engine';
+export { metadataResolver, erc20Handler } from './asset-handlers';
